fix: eliminate cascade re-renders from music player time updates
Convert all MUSIC_PLAYER.read() whole-store subscriptions to field-level
selectors using MUSIC_PLAYER.resolve().field(), preventing UI components
from re-rendering on every current_time tick during playback.

Key changes:
- Layout (routes/mod.rs): field-level reads for is_visible, current_track,
  view_mode instead of MUSIC_PLAYER.read()
- PlayerExpanded: extract seek bar into ExpandedSeekBar child component
  so only the seek bar re-renders on time ticks, not the entire 300-line
  expanded player UI
- 8 card/page components: convert use_memo/use_effect MUSIC_PLAYER.read()
  to resolve().current_track() + resolve().is_playing() field selectors
  (unified_track_card, track_card, radio_card, episode_card,
   external_content_card, radio_station, podcast_episode_detail,
   track_detail, zap_dialog)
- Infinite scroll mobile polling: reduce from 250ms to 1000ms interval
- Remove stale home.rs.bak
