+++ src/components/music/radio_card.rs
@@ -1,6 +1,6 @@
 use crate::components::icons;
 use crate::routes::Route;
-use crate::stores::music_player::{self, MusicTrack, MUSIC_PLAYER};
+use crate::stores::music_player::{self, MusicPlayerStateStoreExt, MusicTrack, MUSIC_PLAYER};
 use crate::utils::radio::{get_ranked_stream_urls, RadioStation};
 use dioxus::prelude::*;
 #[derive(Props, Clone, PartialEq)]
@@ -17,9 +17,10 @@ pub fn RadioCard(props: RadioCardProps) -> Element {
     let station_id = station.coordinate.clone();
     let station_id_for_memo = station_id.clone();
     let is_playing = use_memo(move || {
-        let player_state = MUSIC_PLAYER.read();
-        if let Some(ref current) = player_state.current_track {
-            current.id == station_id_for_memo && player_state.is_playing
+        let store = MUSIC_PLAYER.resolve();
+        let current = store.current_track().cloned();
+        if let Some(ref cur) = current {
+            cur.id == station_id_for_memo && store.is_playing().cloned()
         } else {
             false
         }
