+++ src/components/podcast/episode_card.rs
@@ -15,7 +15,7 @@ use crate::components::icons;
 use crate::routes::Route;
 use crate::services::podcast_index::{Episode as PodcastIndexEpisode, PodcastFeed};
 use crate::services::podcast_rss::{format_duration, RssEpisode, RssPodcast};
-use crate::stores::music_player::{self, MusicTrack};
+use crate::stores::music_player::{self, MusicPlayerStateStoreExt, MusicTrack};
 use crate::stores::nostr_music::TrackSource;
 use crate::utils::podcast::{Person, PodcastEpisode, Soundbite, TranscriptRef, ValueBlock};
 /// Unified podcast episode for display
@@ -335,9 +335,10 @@ pub fn PodcastEpisodeCard(props: PodcastEpisodeCardProps) -> Element {
     let episode = &props.episode;
     let episode_id_for_memo = episode.id.clone();
     let is_playing = use_memo(move || {
-        let player_state = music_player::MUSIC_PLAYER.read();
-        if let Some(ref current) = player_state.current_track {
-            current.id == episode_id_for_memo && player_state.is_playing
+        let store = music_player::MUSIC_PLAYER.resolve();
+        let current = store.current_track().cloned();
+        if let Some(ref cur) = current {
+            cur.id == episode_id_for_memo && store.is_playing().cloned()
         } else {
             false
         }
